Add files via upload
Problem Statement
Multiple emails from customers/clients that are dealt with different teams based on the context. This solution should enable auto-classification of emails based on the context, so the same can be routed to best suited team for further processing.

Technology:

Python
Anaconda platform
Other Considerations:

The data selected should represent sufficient variation to be able to demonstrate classification clearly. Expectation from participants will be to present overall solution with clear focus on characteristics of data and holistic nature of the implementation.

Dataset:

For solving this problem, participants can decide to leverage data available on public forums like Kaggle (preferably from finance domain). But the model should be easy to configure/retrain for similar topics.

Design Considerations:

This model should be easy to deploy to execute either as batch or real time. Focus should also be on making it efficient from resource consumption standpoint and something that can be hosted as containers.

Benefits:

Auto-email classification will enable significant reduction in manual efforts

Solution
Note

The solution mentioned below is what we had implemented for the hackathon in April 2024. Since then, running LLMs on-device has become much easier thanks to Ollama, hence we have updated the project to Ollama instead of using a cloud-hosted model.

Demo Video


Process Flow Diagram

Working
Our solution works as follows:

Email Receiving: We have a root node, which acts as the central email receiver in an organization. It is continuously monitored for new emails through the IMAP protocol (which is done in a continuously running python service). It can be any professional email service like Gmail, Outlook, etc.
Distributed Task Queue: Each email is then sent to a distributed task queue, where it is processed by multiple parallel workers, thus ensuring that the system is scalable and can handle a large number of emails. The system can thus scale horizontally very effectively.
Parsing: The email is parsed and the content is extracted (including attachments!)
Encryption: The content is encrypted using a powerful cypher.
Sending to Server: The encrypted content is sent to the server where it is decrypted.
Large Language Model: The content is passed to our powerful Large Language Model, Mixtal 8x7b which extracts the meaning, context and the sentiment from the email and determines the department/team to which the email should be routed to, including any CCs or BCCs that it deems important.
Summarization: The email is then summarized by a lighter model, and the summary is sent back to the service.
Message Composition: The service then composes a new email with the summary (for quick reference) and the sentiment (e.g. Urgent, Complaint, Neutral) and the original email (along with attachments), and sends it back to the root node.
Routing: The full email is then routed to the respective department/team/person through the SMTP Protocol.
Feedback System: The system also has a feedback system, where in the case of incorrect routing, the user can provide feedback, which is then used to retrain the model.